pyproject.toml
[project]
name = "blogmore"
version = "0.1.0"
description = "A blog-oriented static site generation engine"
readme = "README.md"
authors = [
    { name = "Dave Pearson", email = "davep@davep.org" }
]
requires-python = ">=3.12"
license = "GPL-3.0-or-later"
keywords = [
    "blog",
    "static site generator",
    "markdown",
    "jinja2",
]
dependencies = [
    "jinja2>=3.1.4",
    "markdown>=3.7",
    "python-frontmatter>=1.1.0",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Typing :: Typed",
]

[project.urls]
Homepage = "https://github.com/davep/blogmore"
Repository = "https://github.com/davep/blogmore"
Source = "https://github.com/davep/blogmore"
Issues = "https://github.com/davep/blogmore/issues"

[project.scripts]
blogmore = "blogmore.__main__:main"

[build-system]
requires = ["uv_build>=0.9.4"]
build-backend = "uv_build"

[tool.pyright]
venvPath="."
venv=".venv"
exclude=[".venv"]

[dependency-groups]
dev = [
    "codespell>=2.4.1",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "ruff>=0.14.0",
    "types-markdown>=3.7.0.20250106",
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = [
    # I think try...except...pass reads far better.
    "SIM105",
]

[tool.ruff.lint.pycodestyle]
max-line-length = 120
